apps/api/pyproject.toml
[project]
name = "roborouter-api"
version = "0.1.0"
dependencies = [
  "fastapi==0.114.1",
  "uvicorn[standard]==0.30.6",
  "pydantic-settings==2.4.0",
  "sqlalchemy==2.0.36",
  "psycopg2-binary==2.9.9",
  "minio==7.2.8",
]


